Update LQA response
